adding new connection types (data, control)
